fix: Cast Byte to Integer for decimalPlaces in CurrencyResponse

+++ src/main/java/com/riverflow/service/admin/PackageService.java
@@ -295,7 +295,7 @@ public class PackageService {
                 .code(currency.getCode())
                 .name(currency.getName())
                 .symbol(currency.getSymbol())
-                .decimalPlaces(currency.getDecimalPlaces())
+                .decimalPlaces(currency.getDecimalPlaces().intValue())
                 .isActive(currency.getIsActive())
                 .build();
     }
